fix synx images

+++ files/sbin/lxc-snap.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+set -e
 if [ "x${DEBUG}" != "x" ] ;then
     set -x
 fi
@@ -7,7 +8,9 @@ if [ "x$1" = "xstop" ];then
         service $i stop
     done
 fi
-rm -rfv /tmp/.saltcloud
+if [ -e /tmp/.saltcloud ];then
+    rm -rfv /tmp/.saltcloud
+fi
 for i in /var/cache/salt/{salt-master,salt-minion} /var/cache/mastersalt/{master-master,mastersalt-minion} /etc/.git;do
     if [ -e "${i}" ];then
         rm -rfv "${i}"/*
@@ -15,7 +18,9 @@ for i in /var/cache/salt/{salt-master,salt-minion} /var/cache/mastersalt/{master
 done
 for i in /tmp;do
     if [ -e "${i}" ];then
-        cd "${i}" && find|while read f;do rm -rfv "${f}";done
+        cd "${i}" && find|while read f;do
+            if [ "x${f}" != "x." ];then rm -rfv "${f}";fi
+        done
     fi
 done
 find /etc/*salt/pki/master/minions*/*\
@@ -32,35 +37,38 @@ find /etc/*salt/pki/master/minions*/*\
     /var/log/mastersalt/mastersalt-minion\
     /var/log/salt/salt-master\
     /var/log/salt/salt-minion\
-    | while read fic;do rm -fv "${fic}";done
+    | while read fic;do rm -fv "${fic}" || /bin/true ;done
 if [ -e /var/lib/apt/lists ];then
-    find /var/lib/apt/lists -type f -delete
+    find /var/lib/apt/lists -type f -delete || /bin/true
 fi
 find /srv -name .git|while read f
 do
-    cd "${f}"
+    cd "${f}/.."
     gmarker=".git/curgitpackid"
-    if [ -f $gmarker ];then
+    if [ ! -f $gmarker ];then
         echo 1 > $gmarker
     fi
     gitpackid=$(cat $gmarker)
-    if [ "x$((echo "${gitpackid}"%10))" = "x0" ];then
-        git prune
-        git gc --aggressive
+    if [ "x$((${gitpackid}%10))" = "x0" ];then
+        git prune || /bin/true
+        git gc --aggressive || /bin/true
         echo $((${gitpackid}+1)) > $gmarker
     fi
 done
 find /srv/pillar /srv/mastersalt-pillar /etc/*salt/minion* -type f|while read i
 do
-    sed -i -e "s/master:.*/master: 0.0.0.1/g" "$i"
+    sed -i -re "s/master:.*/master: 0.0.0.1/g" "$i"
 done
 find /etc/shorewall/rules -type f|while read i
 do
-    sed -i -e -re 's/ACCEPT.? +net:?.*fw +-/ACCEPT net fw/g' "$i"
+    sed -i -re "s/ACCEPT.? +net:?.*fw +-/ACCEPT net fw/g" "$i"
 done
 find / -name .bash_history | while read fic;do echo >"${fic}";done
 find /etc/init/*salt* |grep -v override| while read fic
     do
         echo manual > ${fic//.conf}.override
     done
-find /var/cache/apt/archives -name "*deb" -type f -delete 2>/dev/null
+find /var/cache/apt/archives -name "*deb" -type f -delete 2>/dev/null || /bin/true
+if [ -e /etc/ssl/cloud ];then
+    find /etc/ssl/cloud -type -f -delete 2>/dev/null || /bin/true
+fi
